bpf/lib: unconditionally create ipcache bpf map in datapath
In case the opening and creating of the map in userspace fails, it now will be
unconditionally created in the datapath. This will not affect the datapath
because the lookup into the map is only performed when egress policy is enabled
in the datapath for a specific endpoint.

Signed-off by: Ian Vernon <ian@cilium.io>